refactor(architecture): :building_construction: Improves project folder structuring
