+++ app/backend/src/middlewares/deleteTalker.js
