+++ app/backend/src/middlewares/getTalkerById.js
