+++ app/backend/src/middlewares/getTalkers.js
