+++ app/backend/src/middlewares/newTalker.js
