+++ app/backend/src/routes/Login.routes.js
@@ -1,10 +1,10 @@
 const { Router } = require('express');
 const router = Router();
 
-const { createLogin } = require('../../middlewares/createLogin');
+const { createLogin } = require('../middlewares/createLogin');
 const {
   validateEmail,
-  validatePassword } = require('../../validations/validateUser');
+  validatePassword } = require('../validations/validateUser');
 
 router.post('/login', validateEmail, validatePassword, createLogin);
