+++ app/backend/src/routes/Talker.routes.js
@@ -1,18 +1,18 @@
 const { Router } = require('express');
 const router = Router();
 
-const { getTalkers } = require('../../middlewares/getTalkers');
-const { getTalkerById } = require('../../middlewares/getTalkerById');
+const { getTalkers } = require('../middlewares/getTalkers');
+const { getTalkerById } = require('../middlewares/getTalkerById');
 
 const {
   validateToken,
   validateName,
   validateAge,
-  validateTalk } = require('../../validations/validateTalker');
-const { filterByName } = require('../../middlewares/filterByName');
-const { deleteTalker } = require('../../middlewares/deleteTalker');
-const { createTalker } = require('../../middlewares/newTalker');
-const { editTalker } = require('../../middlewares/editTalker');
+  validateTalk } = require('../validations/validateTalker');
+const { filterByName } = require('../middlewares/filterByName');
+const { deleteTalker } = require('../middlewares/deleteTalker');
+const { createTalker } = require('../middlewares/newTalker');
+const { editTalker } = require('../middlewares/editTalker');
 
 
 router.get('/talker', getTalkers);
