+++ app/backend/src/validations/validateTalker.js
